retake sur les derniers bug concernant la mise en ligne

+++ details_mission.php
@@ -24,14 +24,18 @@ $planques = $managerPlanques->getByMission($_GET['id']);
 
 <div class="container">
 
-    <ul class="nav justify-content-end m-3">
-        <li class="nav-item m-1 p-2">
-        <a href="actions/missions/updateMission.php?id=<?= $mission->getMission_id_uuid() ?>" ><img src="https://img.icons8.com/external-kiranshastry-solid-kiranshastry/45/000000/external-edit-interface-kiranshastry-solid-kiranshastry-2.png"/></a>
-        </li>
-        <li class="nav-item m-1 p-2">
-        <a href="actions/missions/deleteMission.php?id=<?= $mission->getMission_id_uuid() ?>" ><img src="https://img.icons8.com/ios-glyphs/45/000000/filled-trash.png"/></a>
-        </li>
-    </ul>
+    <?php if (isset($_SESSION['last_name'])) { ?>
+        <ul class="nav justify-content-end m-3">
+            <li class="nav-item m-1 p-2">
+                <a href="actions/missions/updateMission.php?id=<?= $mission->getMission_id_uuid() ?>"><img src="https://img.icons8.com/external-kiranshastry-solid-kiranshastry/45/000000/external-edit-interface-kiranshastry-solid-kiranshastry-2.png" /></a>
+            </li>
+            <li class="nav-item m-1 p-2">
+                <a href="actions/missions/deleteMission.php?id=<?= $mission->getMission_id_uuid() ?>"><img src="https://img.icons8.com/ios-glyphs/45/000000/filled-trash.png" /></a>
+            </li>
+        </ul>
+    <?php
+    }
+    ?>
 
     <h1 class="text-center m-4">Détail de la mission <?= $mission->getCode_name() ?></h1>
